relations sujet messages user, suppression en fonction de luser ou de ses droits

+++ src/main/java/com/evergreen/web/MessageController.java
@@ -2,6 +2,7 @@ package com.evergreen.web;
 
 import com.evergreen.entities.Message;
 import com.evergreen.entities.Sujet;
+import com.evergreen.entities.UserSession;
 import com.evergreen.service.MessageService;
 import com.evergreen.service.SujetService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,8 @@ public class MessageController {
     private MessageService messageService;
     @Autowired
     private SujetService sujetService;
+    @Autowired
+    private UserSession userSession;
 
     @PostMapping("message")
     public String saveMessage(@RequestParam(name = "ref", defaultValue = "") Long idSujet,
@@ -32,7 +35,9 @@ public class MessageController {
     public String deleteMessage(@RequestParam(name = "ref", defaultValue = "") Long idMessage,
                               Model model) {
         Message message = messageService.getMessage(idMessage).get();
-        messageService.delete(idMessage);
+        if(message.getUser().getId() == userSession.getId() || userSession.isAdministrator()){
+            messageService.delete(idMessage);
+        }
         return "redirect:/sujet?ref=" + message.getSujet().getIdSujet();
     }
 }
